requirements.txt
numpy
opencv-python
scikit-learn
Pillow
matplotlib
scikit-image
torch
torchvision
transformers